Refactor: Modularize application structure
- Separated concerns into core, utils, generation, tools, and orchestration modules.
- idiomatic.py now serves as the main entry point, handling initializations and graph execution.
- Improved code organization and maintainability.
- Updated imports and type hints for the new structure.
